clear documention

+++ main.dart
@@ -5,7 +5,7 @@ void main() {
   List<double> Lantiide = [40, 41, 41.5];
   List<double> Oldlogitude = [23, 24, 26];
   List<double> OldLantiide = [40, 41, 42];
-  const int time_passed = 8000000;
+  const double time_passed = 8000000;
   ai_data cool =
       Ai().main(logitude, Lantiide, Oldlogitude, OldLantiide, time_passed);
   print(cool.AvgSpeed);
